# [IMP] core: Environment._ for translations and LazyTranslate