--- a/addons/hr/models/hr_job.py
+++ b/addons/hr/models/hr_job.py
@@ -47,7 +47,7 @@ class Job(models.Model):
 
     def copy_data(self, default=None):
         vals_list = super().copy_data(default=default)
-        return [dict(vals, name=_("%s (copy)", job.name)) for job, vals in zip(self, vals_list)]
+        return [dict(vals, name=self.env._("%s (copy)", job.name)) for job, vals in zip(self, vals_list)]
 
     def write(self, vals):
         if len(self) == 1: