--- a/addons/hr_holidays/models/hr_leave_accrual_plan.py
+++ b/addons/hr_holidays/models/hr_leave_accrual_plan.py
@@ -139,7 +139,7 @@ class AccrualPlan(models.Model):
 
     def copy_data(self, default=None):
         vals_list = super().copy_data(default=default)
-        return [dict(vals, name=_("%s (copy)", plan.name)) for plan, vals in zip(self, vals_list)]
+        return [dict(vals, name=self.env._("%s (copy)", plan.name)) for plan, vals in zip(self, vals_list)]
 
     @api.ondelete(at_uninstall=False)
     def _prevent_used_plan_unlink(self):