--- a/addons/hr_holidays/models/hr_leave_type.py
+++ b/addons/hr_holidays/models/hr_leave_type.py
@@ -377,7 +377,7 @@ class HolidaysType(models.Model):
 
     def copy_data(self, default=None):
         vals_list = super().copy_data(default=default)
-        return [dict(vals, name=_("%s (copy)", leave_type.name)) for leave_type, vals in zip(self, vals_list)]
+        return [dict(vals, name=self.env._("%s (copy)", leave_type.name)) for leave_type, vals in zip(self, vals_list)]
 
     def action_see_days_allocated(self):
         self.ensure_one()