--- a/addons/l10n_es_pos/__init__.py
+++ b/addons/l10n_es_pos/__init__.py
@@ -1,4 +1,3 @@
-from odoo import _
 from . import models