… addons/account_test/report/report_account_test.py
… addons/l10n_ec_website_sale/controllers/main.py
… addons/l10n_es_edi_tbai/models/l10n_es_edi_tbai_document.py
… addons/l10n_es_edi_tbai/models/res_company.py
… addons/l10n_hu_edi/models/account_tax.py
… addons/l10n_in/models/account_invoice.py
… addons/l10n_in_edi_ewaybill/models/account_edi_format.py
… addons/l10n_in_edi_ewaybill/models/error_codes.py
… addons/l10n_in_ewaybill_stock/tools/ewaybill_api.py
… addons/l10n_pe_website_sale/controllers/main.py
… addons/mail/models/ir_mail_server.py
… addons/mail/models/mail_template.py
… addons/mail/models/mail_tracking_value.py
… addons/mail/tools/parser.py
… addons/mail/wizard/base_partner_merge_automatic_wizard.py
… addons/mail/wizard/mail_activity_schedule.py
… addons/mail/wizard/mail_compose_message.py
… addons/mass_mailing/models/mailing_list.py
… addons/mrp/models/mrp_unbuild.py
… addons/mrp/models/stock_picking.py
… addons/onboarding/models/onboarding_onboarding.py
… addons/onboarding/models/onboarding_onboarding_step.py
… addons/payment/const.py
… addons/payment_mercado_pago/const.py
… addons/phone_validation/tools/phone_validation.py
… addons/product/models/product_tag.py
… addons/project/models/project_project.py
… addons/project/models/project_project_stage.py
… addons/project/models/project_task_type.py
… addons/project_account/models/project_project.py