Python translations need to know the language and the module to find the translated term. Today's `_` function must resolve both of these values from the current stack by using some heuristics.
These heuristics don't always work properly: to find the language we look in the caller's scope to find it in the context or to find the environment from which we can read lang.
For the module, if the function is used outside of an addon, we default to 'base'.

We add a function `Environment._` for translations. That function uses its `lang` property for the language without resorting to inspecting the call stack. You can also pass the module, but we can still resolve it dynamically.

For lazy translations, we create a new `LazyTranslate` factory that you can use as `_lt = LazyTranslate(__name__)` in any addon. This will resolve the addon name once per file from the name of the package without going through `inspect`. You can then use `Environment._` to get the translation or `str()` to resolve the lanugage dynamically.


odoo/enterprise#67528
odoo/documentation#10421
task-4034999

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr