@robodoo 2024-07-29
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/174844.png)](https://mergebot.odoo.com/odoo/odoo/pull/174844)