@kmagusiak 2024-07-29
# Performance improvement

## Env translation
```python
def testme(self, f):
    return f("some string")

%timeit testme(self, _)
%timeit testme(self, self.env._)
TODO

```

## Lazy
```python
# before
%timeit _lt("hello world")
# 45.3 μs ± 201 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

# after
%timeit lt = LazyTranslate('odoo.addons.test')
# 2.63 μs ± 27.3 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
%timeit lt("hello world")
# 4.05 μs ± 50 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```
For 10 translations we go from 450μs to 43μs: 10x improvement.