@kmagusiak 2024-07-31
> `Envrionment` ?

Class `odoo.api.Environment` the instance is available on models as `self.env`