@tde-banana-odoo 2024-07-31
No, it's just that there is a typo in commit header message ^^