@kmagusiak 2024-08-21
> I think the changes might break the current Pylint checks.
> 
> https://github.com/odoo/odoo/blob/2ee1603ab4b9c7f7ad3fb5f77dde77618ddd1e89/odoo/addons/test_lint/tests/_odoo_checker_gettext.py#L56
> 
> I believe that `isinstance(node.func, astroid.Name)` ensures that the function is a simple, unqualified identifier. Meaning it matches `_()`, but not `self.env._()`. Could you add some illegal code on purpose and see if it's still detected by the linter?

Indeed, I fixed the linter and added a test to check if both functions are detected in the same way.