@rco-odoo 2024-09-09
The performance improvement of `_lt("hello world")` does not make much sense.  If you use `_lt` is in the module's toplevel code, it is run once and only for the module part.  The real translation occurs when you call `str(msg)`, which will introspect the stack frame to retrieve the target language.  So overall the actual translation speedup is not that big, is it?