@kmagusiak 2024-09-09
> I am quite skeptical about the claimed performance improvement, since you actually don't take really advantage of it. The only case where it is effective is when no translation is needed (`lang` is empty or `lang="en_US"`). It's only a shortcut for cases that are not common on a non-English database.

Performance improvement is minor (if any), but the goal was to show it is not worse.
Note that I see most of people working in the backoffice using en_US :)

> Also, using `_lt()` for exception messages feels like really bad practice to me, since the translation does not occur where it used to do. Relying on some lazy evaluation that depends on the execution environment is a recipe for bugs IMO.

I feel like lazy translations are misunderstood: they exist so that we can define the translation only once and use it in multiple places in the same file. The right thing to do is to run `env._(some_lazy_translation)` instead of returning them from any module. Using them inside exceptions is also a bad practice for the reasons you give, but I felt it was out of scope of this PR.
Actually `str(lazy_translation)` should log a warning in the future or we could remove lazy translations altogether (if needed, make a helper function that returns a translated string).