@moylop260 2025-01-13
Hi, the following lint detects all the old cases using `_` method:
 - https://github.com/odoo/odoo/pull/193345

And there are a lot of occurrences:
 - https://gist.github.com/moylop260/71311407f4a00f9a942a4e51757b00f6

My question is, what is the plan with all these sentences using the old way `_`?

I mean, it will be migrated to new `self.env._` or `_lt`?